requirements.txt
python-telegram-bot==21.7
PyPDF2==3.0.1
pdfplumber==0.10.3
google-generativeai==0.3.0
requests==2.31.0
python-dotenv==1.0.0
pymupdf==1.23.8